Pass the buffer data as 2 arguments rather than a struct
This continues the path of trying to avoid structs because they cause
JNA slowness.

Results (min, median, max): 292ms 329ms 366ms

This seems like a good stopping point.  It's still slower than the
hand-written times of 182ms, but that seems fine.  I feel very
comfortable telling people that the overhead of the FFI call itself is
a bit under 2x the cost of a hand-written FFI.  Presumably the work
inside the call is going to be much more.