Added a generate items maintenance function
I'm hoping that this will help us trigger a crash in places.  By
generating new items, we ensure the next maintenance pass will have
something to do -- especially the optimize and checkpoint part.  This
hopefully increases the chance that we trigger the EXC_BAD_ACCESS error.

I made it generate 1000 extra items, feel free to tune this.  Maybe we
should also delete items? My intuition is that it's best to just let the
DB keep growing.