Remote settings persistence support
Added the `RemoteSettingsService` and `RemoteSettingsClient` types.
These represent the shape of the new API that we're hoping to implement.
The goal is to match the API of the official client more closely. In
particular, the new types have different semantics for getting/syncing
records that allow clients to avoid hitting the network when getting
settings values during startup.