Different way to match weather suggestions
Replace `filter_map_chunked` with a way to iterate over chunks in a
string.

Implement a depth-first-search in weather to find the suggestions.  This
is simple enough that we don't need any generic code, if we want to do
this for a different suggestion type we can follow the same pattern
rather than using shared code.

Fixed the `ny rochester` case.